refactor(auth): update core auth folder to utils for better naming convention
