+++ apps/server/src/modules/auth/features/oauth/google/google.module.ts
@@ -2,8 +2,8 @@ import { googleAuthConfig } from "@/modules/auth/config/google-auth.config";
 
 import { userService } from "@/modules/user/user.module";
 import { cacheService } from "@/modules/utils/cache/cache.module";
-import { jwtService } from "@modules/auth/core/services/jwt/jwt.module";
 import { refreshTokenService } from "@modules/auth/features/refresh-token/refresh-token.module";
+import { jwtService } from "@modules/auth/utils/services/jwt/jwt.module";
 
 import { GoogleController } from "./google.controller";
 import { GoogleAuthService } from "./google.service";
