+++ apps/server/src/modules/auth/features/refresh-token/refresh-token.module.ts
@@ -1,6 +1,6 @@
 import prismaClient from "@config/database.config";
 
-import { jwtService } from "@modules/auth/core/services/jwt/jwt.module";
+import { jwtService } from "@modules/auth/utils/services/jwt/jwt.module";
 
 import { RefreshTokenController } from "./refresh-token.controller";
