+++ apps/server/src/modules/auth/features/refresh-token/refresh-token.service.ts
@@ -8,7 +8,7 @@ import { ERROR_MESSAGES } from "@constants/error-messages.contants";
 import { stringToDate } from "@utils/convert-string-to-date";
 import logger from "@utils/logger";
 
-import { IJwtService } from "@modules/auth/core/services/jwt/jwt.types";
+import { IJwtService } from "@modules/auth/utils/services/jwt/jwt.types";
 
 import type { AuthResponseDto, RefreshDto } from "@modules/auth/dtos";
