+++ apps/server/src/modules/auth/features/verify-email/verify-email.module.ts
@@ -1,6 +1,6 @@
 import prismaClient from "@config/database.config";
 
-import { jwtService } from "@modules/auth/core/services/jwt/jwt.module";
+import { jwtService } from "@modules/auth/utils/services/jwt/jwt.module";
 import { userService } from "@modules/user/user.module";
 
 import { VerifyEmailController } from "./verify-email.controller";
