+++ apps/server/src/modules/auth/features/verify-email/verify-email.service.ts
@@ -8,7 +8,7 @@ import { BadRequestError, ForbiddenError, NotFoundError } from "@/errors";
 
 import { stringToDate } from "@utils/convert-string-to-date";
 
-import type { IJwtService } from "@modules/auth/core/services/jwt/jwt.types";
+import type { IJwtService } from "@modules/auth/utils/services/jwt/jwt.types";
 import type { IUserService } from "@modules/user/user.types";
 
 import type {
