+++ apps/server/src/modules/auth/utils/services/jwt/__test__/jwt.service.test.ts
