+++ apps/server/src/modules/auth/utils/services/jwt/dtos/sign-token.dto.ts
