+++ apps/server/src/modules/auth/utils/services/jwt/jwt.types.ts
