+++ apps/server/src/modules/user/user.module.ts
@@ -1,6 +1,6 @@
 import prismaClient from "@config/database.config";
 
-import { passwordHasherService } from "@modules/auth/core/services/password-hasher/password-hasher.module";
+import { passwordHasherService } from "@modules/auth/utils/services/password-hasher/password-hasher.module";
 
 import { UserController } from "./user.controller";
 import { UserRepository } from "./user.repository";
